requirements.txt
Django>=3.1.0,<=3.1.9
djangorestframework>=3.12.4,<=3.13.0
psycopg2>=2.7.6,<=2.8.6

flake8>=3.8.0,<=3.9.0
